A5: fix compiler warning

+++ sets/set5/A5/Quadrilateral.cpp
@@ -9,5 +9,6 @@
 Quadrilateral::Quadrilateral() : Polygon()
 {
     this->numVertices = 4;
-    this->vertices = new Coordinate[numVertices];
+    this->vertices = new Coordinate[static_cast<std::size_t>(numVertices)];
+    ;
 }
